+++ buildSrc/src/main/java/MyApp.kt
@@ -2,6 +2,6 @@
 
 object MyApp {
     const val appId = "com.sysadmindoc.iosicons"
-    const val version = 9
-    const val versionName = "1.1.7"
+    const val version = 10
+    const val versionName = "1.1.8"
 }